move `printCandidate` normalization tests to core
